Add files via upload
The management wants to understand the conditions of the sales activities and gain insights into the various trends happening in the Bike Sales volume between 2016 and 2018. They also want to know the revenues per region and per store. Excel Pivot tables and a dashboard was created to visualize these results.